Updated Compile On Save, and Empty Functions Allowed
Previously, when the source file would be compiled with the 'Compile on source changed' checkbox enabled, it would always overwrite the clipboard. Now, the 'asm' output is automatically updated when the source file is changed, but it does not overwrite the clipboard.

Also, empty functions would cause compiler errors. Empty 'main()' and user-defined functions can now be defined.